// moca_LP/src/types/lp.ts
import type { CommonResponse, CursorBasedResponse } from "./common";

export type Tag = {
  id: number;
  name: string;
};

export type Likes = {
  id: number;
  userId: number;
  lpId: number;
}

export type LP = {
  id: number;
  title: string;
  content: string;
  thumbnail: string;
  published: boolean;
  authorld: number;
  createdAt: Date;
  updatedAt: Date;
  tags: Tag[];
  likes: Likes[];
};

export type RequestLpDto = {
  lpId: number;
};

export type ResponseLpDto = CommonResponse<LP>;

export type ResponseLPListDto = CursorBasedResponse<LP[]>

export type ResponseLikeLpDto = CommonResponse<{
  id: number;
  userId: number;
  lpId: number;
}>

export type CommentAuthor = {
  id: number;
  name: string;
  email: string;
  bio: string | null;
  avatar: string | null;
  createdAt: string;
  updatedAt: string;
};

export type Comments = {
  id: number;
  content: string;
  lpId: number;
  authorId: number;
  createdAt: string;
  updatedAt: string;
  author: CommentAuthor;
}


export type GetCommentParams = {
  lpId: number;
  cursor: number;
  limit: number;
  order: 'asc' | 'desc';
}

export type ResponseCommentListDto = CursorBasedResponse<Comments[]>;

export type CreateLpDto = {
  title: string;
  content: string;
  thumbnail: string;
  tags?: string[];
  published?: boolean;
}